+++ malaria/app.py
@@ -7,48 +7,7 @@ import numpy as np
 from tensorflow.keras.applications.imagenet_utils import preprocess_input, decode_predictions
 from tensorflow.keras.models import load_model
 from tensorflow.keras.preprocessing import image
-from flask import Flask, redirect, url_for, request, render_template
 from werkzeug.utils import secure_filename
-
-# Define a flask app
-app = Flask(__name__)
-MODEL_PATH ='model.h5'
-model = load_model(MODEL_PATH)
-def model_predict(img_path, model):
-    img = image.load_img(img_path, target_size=(224, 224))
-    x = image.img_to_array(img)
-    x=x/255
-    x = np.expand_dims(x, axis=0)
-    x = preprocess_input(x)
-
-    preds = model.predict(x)
-    preds=np.argmax(preds, axis=1)
-    print(preds)
-    return preds
-
-@app.route('/', methods=['GET'])
-def index():
-    return render_template('index.html')
-
-@app.route('/predict', methods=['GET', 'POST'])
-def upload():
-    output = "Invalid request method"
-    if request.method == 'POST':
-        # Get the file from post request
-        f = request.files['file']
-        basepath = os.path.dirname(__file__)
-        file_path = os.path.join(
-            basepath, 'uploads', secure_filename(f.filename))
-        f.save(file_path)
-
-        # Make prediction
-        preds = model_predict(file_path, model)
-
-        class_names =  ["The Person is Infected With Malaria.","The Person is not Infected With Malaria."]
-        output = class_names[preds[0]]
-      
-    return render_template('index.html', pred = output)
+from flask import Flask, redirect, url_for, request, render_template
 
 
-if __name__ == '__main__':
-    app.run(debug=True)
