--- malaria/templates/base.html
@@ -1,33 +1,0 @@
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Malaria</title>
-    <link href="https://cdn.bootcss.com/bootstrap/4.0.0/css/bootstrap.min.css" rel="stylesheet">
-    <script src="https://cdn.bootcss.com/popper.js/1.12.9/umd/popper.min.js"></script>
-    <script src="https://cdn.bootcss.com/jquery/3.3.1/jquery.min.js"></script>
-    <script src="https://cdn.bootcss.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
-    <link href="{{ url_for('static', filename='css/main.css') }}" rel="stylesheet">      
-</head>
-
-<body>
-    <nav class="navbar navbar-dark bg-dark">
-        <div class="container">
-            <a class="navbar-brand" href="#">Upload Your Blood Sample Report for Malaria Dtection</a>
-        </div>
-    </nav>
-    <div class="container">
-        <div id="content" style="margin-top:2em">{% block content %}{% endblock %}</div>
-    </div>
-    <div class="container">
-        <div id="content" style="margin-top:2em">{{ pred }}</div>
-    </div>
-</body>
-
-<footer>
-    <script src="{{ url_for('static', filename='js/main.js') }}" type="text/javascript"></script>    
-</footer>
-
-</html>
