--- malaria/templates/index.html
@@ -1,23 +1,0 @@
-{% extends "base.html" %} 
-{% block content %}
-
-<h2>Image Classifier</h2>
-
-<div>
-    <form id="upload-file" method="post" enctype="multipart/form-data" action="{{ url_for('upload') }}" >
-        <label for="imageUpload" class="upload-label">
-            Choose...
-        </label>
-        <input type="file" name="file" id="imageUpload" accept=".png, .jpg, .jpeg">
-        <div class="image-section" style="display:none;">
-        <div class="img-preview">
-            <div id="imagePreview">
-            </div>
-        </div>
-        <div>
-            <input type="submit" class="btn btn-primary btn-lg " id="btn-predict" value="Predict!"></input>
-        </div>
-    </div>
-    </form>
-</div>
-{% endblock %}
