+++ templates/index.html
@@ -14,6 +14,11 @@
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
+    <link href="https://cdn.bootcss.com/bootstrap/4.0.0/css/bootstrap.min.css" rel="stylesheet">
+    <script src="https://cdn.bootcss.com/popper.js/1.12.9/umd/popper.min.js"></script>
+    <script src="https://cdn.bootcss.com/jquery/3.3.1/jquery.min.js"></script>
+    <script src="https://cdn.bootcss.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
+    <link href="static/css/main.css" rel="stylesheet"> 
   
   <style>
     html, body{ height:100%; margin:0; }
@@ -48,6 +53,9 @@ footer{
              <li class="nav-item active">
                 <a class="nav-link" href="{{ url_for('kidney') }}">Kidney</a>
               </li>
+             <li class="nav-item active">
+                <a class="nav-link" href="{{ url_for('malaria') }}">Malaria</a>
+              </li>
            </ul>
          </div>
        </nav>
@@ -70,5 +78,6 @@ footer{
     </ul>
   </center>
   </footer>
+  <script src="static/js/main.js" type="text/javascript"></script>
 </body>
 </html>
