+++ templates/malaria.html
@@ -1,0 +1,29 @@
+{% extends "index.html" %} 
+{% block content %}
+
+<h2>Malaria Detector</h2>
+    <nav class="navbar navbar-dark bg-dark" style="text-align: center">
+        <div style="color:#ffffff">
+            Upload your Blood Cell Sample Image for detection
+        </div>
+    </nav>
+<center>
+<div>
+    <form id="upload-file" method="post" enctype="multipart/form-data" action="{{ url_for('predictMalaria') }}" >
+        <br/><br/><label for="imageUpload" class="upload-label">
+            Choose...
+        </label>
+        <input type="file" name="file" id="imageUpload" accept=".png, .jpg, .jpeg">
+        <div class="image-section" style="display:none;">
+        <div class="img-preview">
+            <div id="imagePreview">
+            </div>
+        </div>
+        <div>
+            <input type="submit" class="btn btn-primary btn-lg " id="btn-predict" value="Predict!"></input>
+        </div>
+    </div>
+    </form>
+</div>
+</center>
+{% endblock %}
