+++ templates/predictKidney.html
@@ -1,0 +1,19 @@
+{% extends 'index.html' %}
+{% block content %}
+    <div class="row"  style="margin-bottom: 477px;">
+        <div class="col-md-3"></div>
+        <div class="col-md-6">
+            {% if pred == 1 %}
+                <div class="card card-body alert alert-danger"><center>Unfortunately our algorithm predicted possibility of a kidney disease, please consult a doctor.</center></div>
+            {% else %}
+                <div class="card card-body alert alert-success"><center>No possible disease predicted</center></div>
+            {% endif %}
+            <div class="row">
+                <div class="col-md-4"></div>
+                <div class="col-md-4"><a href="#" class="btn btn-block btn-primary">Back to Home</a></div>
+                <div class="col-md-4"></div>
+            </div>
+        </div>
+        <div class="col-md-3"></div>
+    </div>
+{% endblock %}
