+++ templates/predictMalaria.html
@@ -1,0 +1,10 @@
+{% extends 'index.html' %}
+{% block content %}
+    <div class="row"  style="margin-bottom: 477px;">
+        <div class="col-md-3"></div>
+        <div class="col-md-6">
+                <div class="card card-body alert alert-danger"><center>{{ pred }}</center></div>
+        </div>
+        <div class="col-md-3"></div>
+    </div>
+{% endblock %}
